added static_casts where needed

+++ squantorLibEmbeddedTests/src/test_rmw.cpp
@@ -21,47 +21,47 @@ MINUNIT_ADD(testReadModifyWrite8bit, NULL, NULL) {
   // test AND
   testSource = 0xAA;
   testDestination = 0xCC;
-  util::readModifyWrite(&testDestination, &testSource, 0x0F, 0, util::bitblitOperation::OP_AND);
+  util::readModifyWrite(&testDestination, &testSource, static_cast<uint8_t>(0x0F), 0, util::bitblitOperation::OP_AND);
   minUnitCheck(testDestination == 0xC8);
   // test MOV
   testSource = 0xF;
   testDestination = 0x00;
-  util::readModifyWrite(&testDestination, &testSource, 0x3C, 2, util::bitblitOperation::OP_MOV);
+  util::readModifyWrite(&testDestination, &testSource, static_cast<uint8_t>(0x3C), 2, util::bitblitOperation::OP_MOV);
   minUnitCheck(testDestination == 0x3C);
   testSource = 0xF0;
   testDestination = 0x00;
-  util::readModifyWrite(&testDestination, &testSource, 0x3C, -2, util::bitblitOperation::OP_MOV);
+  util::readModifyWrite(&testDestination, &testSource, static_cast<uint8_t>(0x3C), -2, util::bitblitOperation::OP_MOV);
   minUnitCheck(testDestination == 0x3C);
   testSource = 0x0F;
   testDestination = 0x01;
-  util::readModifyWrite(&testDestination, &testSource, 0x3C, 2, util::bitblitOperation::OP_MOV);
+  util::readModifyWrite(&testDestination, &testSource, static_cast<uint8_t>(0x3C), 2, util::bitblitOperation::OP_MOV);
   minUnitCheck(testDestination == 0x3D);
   testSource = 0x0F;
   testDestination = 0x91;
-  util::readModifyWrite(&testDestination, &testSource, 0x3C, 2, util::bitblitOperation::OP_MOV);
+  util::readModifyWrite(&testDestination, &testSource, static_cast<uint8_t>(0x3C), 2, util::bitblitOperation::OP_MOV);
   minUnitCheck(testDestination == 0xBD);
   testSource = 0x02;
   testDestination = 0x81;
-  util::readModifyWrite(&testDestination, &testSource, 0x03, 0, util::bitblitOperation::OP_MOV);
+  util::readModifyWrite(&testDestination, &testSource, static_cast<uint8_t>(0x03), 0, util::bitblitOperation::OP_MOV);
   minUnitCheck(testDestination == 0x82);
   // test NOT
   testSource = 0xF;
   testDestination = 0xFF;
-  util::readModifyWrite(&testDestination, &testSource, 0x3C, 2, util::bitblitOperation::OP_NOT);
+  util::readModifyWrite(&testDestination, &testSource, static_cast<uint8_t>(0x3C), 2, util::bitblitOperation::OP_NOT);
   minUnitCheck(testDestination == 0xC3);
   testSource = 0xF0;
   testDestination = 0x0F;
-  util::readModifyWrite(&testDestination, &testSource, 0x3C, -2, util::bitblitOperation::OP_NOT);
+  util::readModifyWrite(&testDestination, &testSource, static_cast<uint8_t>(0x3C), -2, util::bitblitOperation::OP_NOT);
   minUnitCheck(testDestination == 0x03);
   // test OR
   testSource = 0xAA;
   testDestination = 0xCC;
-  util::readModifyWrite(&testDestination, &testSource, 0xF0, 4, util::bitblitOperation::OP_OR);
+  util::readModifyWrite(&testDestination, &testSource, static_cast<uint8_t>(0xF0), 4, util::bitblitOperation::OP_OR);
   minUnitCheck(testDestination == 0xEC);
   // test XOR
   testSource = 0xAA;
   testDestination = 0xCC;
-  util::readModifyWrite(&testDestination, &testSource, 0xF0, 4, util::bitblitOperation::OP_XOR);
+  util::readModifyWrite(&testDestination, &testSource, static_cast<uint8_t>(0xF0), 4, util::bitblitOperation::OP_XOR);
   minUnitCheck(testDestination == 0x6C);
 }
 
@@ -71,7 +71,7 @@ MINUNIT_ADD(testReadModifyWrite16bit, NULL, NULL) {
   // test MOV
   testSource = 0xFF;
   testDestination = 0x0000;
-  util::readModifyWrite(&testDestination, &testSource, 0x3C00, 8, util::bitblitOperation::OP_MOV);
+  util::readModifyWrite(&testDestination, &testSource, static_cast<uint16_t>(0x3C00), 8, util::bitblitOperation::OP_MOV);
   minUnitCheck(testDestination == 0x3C00);
   minUnitPass();
   // TODO the rest
@@ -83,7 +83,7 @@ MINUNIT_ADD(testReadModifyWrite32bit, NULL, NULL) {
   // test MOV
   testSource = 0xFF;
   testDestination = 0x00000000;
-  util::readModifyWrite(&testDestination, &testSource, 0x0003C000, 12, util::bitblitOperation::OP_MOV);
+  util::readModifyWrite(&testDestination, &testSource, static_cast<uint32_t>(0x0003C000), 12, util::bitblitOperation::OP_MOV);
   minUnitCheck(testDestination == 0x0003C000);
   // TODO the rest
 }
